refactor(es/common): move Core to Kafka
