+++ modules/evaluation-server/src/Infrastructure/Kafka/KafkaMessageProducer.Log.cs
@@ -1,6 +1,6 @@
 using Microsoft.Extensions.Logging;
 
-namespace Infrastructure.Core;
+namespace Infrastructure.Kafka;
 
 public partial class KafkaMessageProducer
 {
