+++ modules/evaluation-server/src/Infrastructure/Kafka/KafkaMessageProducer.cs
@@ -5,7 +5,7 @@ using Domain.Core;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 
-namespace Infrastructure.Core;
+namespace Infrastructure.Kafka;
 
 public partial class KafkaMessageProducer : IMessageProducer
 {
